% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_tau_mf}
\alias{calculate_tau_mf}
\title{#' @keywords internal
calculate_tau <- function(covariate, w1, w2, theta, m, K) {
    phi_var <- calculate_phi(w1, w2, K)
    covariate <- c(rep(NA, times = K), covariate)
    tau <- exp(sapply(c((K + 1):length(covariate)),
                         FUN = sum_tau,
                         m = m,
                         theta = theta,
                         phivar = phi_var,
                         covariate = covariate,
                         K = K))
    tau
}}
\usage{
calculate_tau_mf(df, x, low.freq, w1, w2, theta, m, K)
}
\description{
#' @keywords internal
calculate_tau <- function(covariate, w1, w2, theta, m, K) {
    phi_var <- calculate_phi(w1, w2, K)
    covariate <- c(rep(NA, times = K), covariate)
    tau <- exp(sapply(c((K + 1):length(covariate)),
                         FUN = sum_tau,
                         m = m,
                         theta = theta,
                         phivar = phi_var,
                         covariate = covariate,
                         K = K))
    tau
}
}
\keyword{internal}
