% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_financial}
\alias{df_financial}
\title{Stock returns and financial conditions.}
\format{
A data frame with 11,306 rows and 5 variables:
\describe{
  \item{date}{date}
  \item{return}{daily S&P 500 log returns times 100}
  \item{rv}{5-minute realized variances}
  \item{week}{a dummy for each year/week combination}
  \item{nfci}{National Financial Conditions Index}
}
}
\source{
\url{https://github.com/onnokleen/mfGARCH/}

\url{https://fred.stlouisfed.org/series/NFCI}
}
\usage{
df_financial
}
\description{
A dataset containing the S&P 500 stock returns and the NFCI
}
\keyword{datasets}
