% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_mfgarch}
\alias{df_mfgarch}
\title{Mixed-frequency data set.}
\format{
A data frame with 11,938 rows and 11 variables:
\describe{
  \item{date}{date}
  \item{return}{daily S&P 500 log returns times 100}
  \item{open_close}{open-close returns}
  \item{rv}{5-minute realized variances}
  \item{vix}{Cboe VIX}
  \item{year_week}{a dummy for each year/week combination}
  \item{dhousing}{changes in housing starts}
  \item{dindpro}{changes in industrial production}
  \item{nai}{NAI}
  \item{nfci}{National Financial Conditions Index}
  \item{year_month}{a dummy for each year/month combination}
}
}
\source{
\url{https://github.com/onnokleen/mfGARCH/}

\url{https://fred.stlouisfed.org}
}
\usage{
df_mfgarch
}
\description{
A dataset containing the S&P 500 stock returns, realized variances and macroeconomic variables
}
\keyword{datasets}
