% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mfgarch.R
\name{fit_mfgarch}
\alias{fit_mfgarch}
\title{Calculates weights as described in}
\usage{
fit_mfgarch(df, y, x, K, low.freq = "Date", var.ratio.freq = NULL)
}
\arguments{
\item{df, }{y, x, K , low.freq, var.ratio.freq.}
}
\description{
This function allows you to express your love of cats.
}
\keyword{fit_mfgarch}

