% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_weighting_scheme}
\alias{plot_weighting_scheme}
\title{This function plots the weighting scheme of an estimated GARCH-MIDAS model}
\usage{
plot_weighting_scheme(x)
}
\arguments{
\item{x}{mfGARCH object obtained by fit_mfgarch}
}
\description{
This function plots the weighting scheme of an estimated GARCH-MIDAS model
}
