% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mfgarch.R
\name{simulate_mfgarch}
\alias{simulate_mfgarch}
\title{This function estimates a multiplicative mixed-frequency GARCH model}
\usage{
simulate_mfgarch(n.days, mu, alpha, beta, gamma, m, theta, w1 = 1, w2, K, psi,
  sigma.psi, low.freq = 1, student.t = NULL)
}
\description{
This function estimates a multiplicative mixed-frequency GARCH model
}
\keyword{simulate_mfgarch}
