% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mfgarch.R
\name{simulate_mfgarch_rvol_dependent}
\alias{simulate_mfgarch_rvol_dependent}
\title{Simulate a GARCH-MIDAS similar to Wang/Ghysels with lagged RVol as covariate}
\usage{
simulate_mfgarch_rvol_dependent(n.days, mu, alpha, beta, gamma, m, theta,
  w1 = 1, w2, K, n.intraday = 288)
}
\arguments{
\item{n.days}{number of days}

\item{mu}{mu}

\item{alpha}{alpha}

\item{beta}{beta}

\item{gamma}{gamma}

\item{m}{m}

\item{theta}{theta}

\item{w1}{w1}

\item{w2}{w2}

\item{K}{K}

\item{n.intraday}{number of maximum intraday returns, default 288}
}
\description{
Simulate a GARCH-MIDAS similar to Wang/Ghysels with lagged RVol as covariate
}
