% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_financial}
\alias{df_financial}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 11,644 rows and 21 variables:
\describe{
  \item{date}{date}
  \item{return}{daily S&P 500 log returns times 100}
  \item{nfci}{daily S&P 500 log returns times 100}
  \item{vix_sqrt_252}{de-annualized VIX}
  \item{year_week}{a dummy for each year/week combination}
}}
\source{
\url{https://github.com/onnokleen/mfGARCH/}
}
\usage{
df_financial
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
