% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mfgarch.R
\name{fit_mfgarch}
\alias{fit_mfgarch}
\title{This function estimates a multiplicative mixed-frequency GARCH model}
\usage{
fit_mfgarch(df, y, x, K, low_freq = "Date", var_ratio_freq = NULL)
}
\arguments{
\item{df, }{y, x, K , low_freq, var_ratio_freq.}
}
\description{
This function estimates a multiplicative mixed-frequency GARCH model
}
\keyword{fit_mfgarch}

