% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mfgarch.R
\name{fit_mfgarch}
\alias{fit_mfgarch}
\title{This function estimates a multiplicative mixed-frequency GARCH model}
\usage{
fit_mfgarch(df, y, x, K, low.freq = "date", var.ratio.freq = NULL)
}
\arguments{
\item{df, }{y, x, K , low.freq, var.ratio.freq.}
}
\description{
This function estimates a multiplicative mixed-frequency GARCH model
}
\keyword{fit_mfgarch}
