% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mfgarch.R
\name{fit_mfgarch}
\alias{fit_mfgarch}
\title{This function estimates a multiplicative mixed-frequency GARCH model. For the sake of numerical stability, it is best to multiply log returns by 100.}
\usage{
fit_mfgarch(
  data,
  y,
  x = NULL,
  K = NULL,
  low.freq = "date",
  var.ratio.freq = NULL,
  gamma = TRUE,
  weighting = "beta.restricted",
  x.two = NULL,
  K.two = NULL,
  low.freq.two = NULL,
  weighting.two = NULL,
  multi.start = FALSE,
  control = list(par.start = NULL)
)
}
\arguments{
\item{data}{data frame containing a column named date of type 'Date'.}

\item{y}{name of high frequency dependent variable in df.}

\item{x}{covariate employed in mfGARCH.}

\item{K}{an integer specifying lag length K in the long-term component.}

\item{low.freq}{a string of the low frequency variable in the df.}

\item{var.ratio.freq}{specify a frequency column on which the variance ratio should be calculated.}

\item{gamma}{if TRUE, an asymmetric GJR-GARCH is used as the short-term component. If FALSE, a simple GARCH(1,1) is employed.}

\item{weighting}{specifies the weighting scheme employed in the long-term component. Options are "beta.restricted" (default) or "beta.unrestricted"}

\item{x.two}{optional second covariate}

\item{K.two}{lag lgenth of optional second covariate}

\item{low.freq.two}{low frequency of optional second covariate}

\item{weighting.two}{specifies the weighting scheme employed in the optional second long-term component. Currently, the only option is "beta.restricted"}

\item{multi.start}{if TRUE, optimization is carried out with multiple starting values}

\item{control}{a list}
}
\value{
A list of class mfGARCH with letters and numbers.
\itemize{
\item par - vector of estimated parameters
\item rob.std.err - sandwich/HAC-type standard errors
\item broom.mgarch - a broom-like data.frame with entries
1) estimate: column of estimated parameters
2) rob.std.err - sandwich/HAC-type standard errors
3) p.value - p-values derived from sandwich/HAC-type standard errors
4) opg.std.err - Bollerslev-Wooldrige/OPG standard errors for GARCH processes
5) opg.p.value - corresponding alternative p-values
\item tau - fitted long-term component
\item g - fitted short-term component
\item df.fitted - data frame with fitted values and residuals
\item K - chosen lag-length in the long-term component
\item weighting.scheme - chosen weighting scheme
\item llh - log-likelihood value at estimated parameter vector
\item bic - corresponding BIC value
\item y - dependent variable y
\item optim - output of the optimization routine
\item K.two - lag-lenth of x.two if two covariates are employed
\item weighting.scheme.two - chosen weighting scheme of x.two (if K.two != NULL)
\item tau.forecast - one-step ahead forecast of the long-term component
\item variance.ratio - calculated variance ratio
\item est.weighting - estimated weighting scheme
\item est.weighting.two - estimated weighting scheme of x.two (if K.two != NULL)
}
}
\description{
This function estimates a multiplicative mixed-frequency GARCH model. For the sake of numerical stability, it is best to multiply log returns by 100.
}
\examples{
\dontrun{
fit_mfgarch(data = df_financial, y = "return", x = "nfci", low.freq = "week", K = 52)
fit_mfgarch(data = df_mfgarch, y = "return", x = "nfci", low.freq = "year_week", K = 52,
x.two = "dindpro", K.two = 12, low.freq.two = "year_month", weighting.two = "beta.restricted")
}
}
\keyword{fit_mfgarch}
