% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mfgarch.R
\name{simulate_mfgarch}
\alias{simulate_mfgarch}
\title{This function simulates a GARCH-MIDAS model. Innovations can follow a standard normal or student-t distribution.}
\usage{
simulate_mfgarch(
  n.days,
  mu,
  alpha,
  beta,
  gamma,
  m,
  theta,
  w1 = 1,
  w2,
  K,
  psi,
  sigma.psi,
  low.freq = 1,
  n.intraday = 288,
  student.t = NULL,
  corr = 0
)
}
\arguments{
\item{n.days}{number of days}

\item{mu}{mu}

\item{alpha}{alpha}

\item{beta}{beta}

\item{gamma}{gamma}

\item{m}{m}

\item{theta}{theta}

\item{w1}{w1}

\item{w2}{w2}

\item{K}{K}

\item{psi}{psi}

\item{sigma.psi}{sigma.psi}

\item{low.freq}{number of days per low-frequency period}

\item{n.intraday}{number of maximum intraday returns}

\item{student.t}{either NULL or degrees of freedom}

\item{corr}{correlation between innovations (should only be used for daily tau)}
}
\description{
This function simulates a GARCH-MIDAS model. Innovations can follow a standard normal or student-t distribution.
}
\examples{
simulate_mfgarch(n.days = 200, mu = 0, alpha = 0.06, beta = 0.92, gamma = 0, m = 0,
theta = 0.1, w1 = 1, w2 = 3, K = 12, psi = 0.98, sigma.psi = 0.1, low.freq = 10)
}
\keyword{simulate_mfgarch}
